fix some warnings and minor things like naming and XCode configs

+++ Audio-Levels/Audio-Levels/TBAudioLevelMeter.m
@@ -26,7 +26,7 @@
     return self;
 }
 
-- (void)setLevel:(float)level
+- (void)setLevel:(double)level
 {
     _level = level;
     [self setNeedsDisplay];
